% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.models_PACKAGE.R
\docType{package}
\name{classic.models-package}
\alias{classic.models}
\alias{classic.models-package}
\title{classic.models: sample data and methods}
\description{
This package introduces includes datasets from the Classic Models database
for the Analytics Designer from OpenText. The goal of this package is to
provide a dataset for learning to integrate R scripts with the BIRT report design.
}
\details{
To learn more about classic.models, start with the vignettes:
`browseVignettes(package = "classic.models")`
}
\author{
\strong{Maintainer}: Shawn Giese \email{sgiese@opentext.com}

}
