% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihub_functions.r
\name{set_connection_info}
\alias{set_connection_info}
\title{Collecting information to connect to your server}
\usage{
set_connection_info()
}
\value{
A file containing the connection information
}
\description{
The default Information Hub connection is to http://localhost
with a user name of 'administrator'. If your server is different
use this function to collect the new information. This updates
a file named ihub_connect.rdata to hold this information for the
other methods to use.
}
