% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihub_functions.r
\name{upload2ihub}
\alias{upload2ihub}
\title{Upload a file to Information Hub}
\usage{
upload2ihub(filename, folder, overwrite)
}
\arguments{
\item{filename}{the name and file extension}

\item{folder}{the path on ihub volume to store file like /Resources/}

\item{overwrite}{true to overwrite existing file}
}
\value{
Response code from the Information Hub server
}
\description{
Upload a file from the R workspace to a location on
Information Hub volume.
}
\examples{
upload2ihub('readme.md','/Resources/','')

}
